MVPLAB_VOICELAB/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.3.7
python-dotenv==1.0.1
deepgram-sdk==3.2.0
pydantic>=2.10.0
websockets==12.0
httpx==0.25.2
python-multipart==0.0.9
pydantic-settings==2.2.1
